Mixin cleanup (#74)
* fix "craft if missing" from ae2wtlib

* clean up mixins
