+++ src/main/java/net/neganote/monilabs/mixin/microverses/ShaderLoaderMixin.java
@@ -1,4 +1,4 @@
-package net.neganote.monilabs.mixin;
+package net.neganote.monilabs.mixin.microverses;
 
 import net.minecraft.resources.ResourceLocation;
 import net.neganote.monilabs.client.render.BlackHoleRendererHelpers;
